unblock double_swap
git-svn-id: https://svn.mcs.anl.gov/repos/NEKCEM/trunk@626 df63514b-fc50-0410-8a93-94e60f5a2dad

+++ src/coIO_nekcem_read.c
@@ -435,11 +435,11 @@ void read3dcells4_( int *eConnect, int *numElems, int *numCells, int *numNodes)
 }
 
 #ifdef UPCASE
-void READFIELD4(int *fldid, double *vals, int *numNodes)
+void READFIELD4(int *fldid, float *vals, int *numNodes)
 #elif  IBM
-void readfield4(int *fldid, double *vals, int *numNodes)
+void readfield4(int *fldid, float *vals, int *numNodes)
 #else
-void readfield4_(int *fldid, double *vals, int *numNodes)
+void readfield4_(int *fldid, float *vals, int *numNodes)
 #endif
 {
 #ifdef MPI
